Fix couple issues from code analysis

+++ picoquic/picoquic_internal.h
@@ -439,7 +439,7 @@ typedef struct st_picoquic_stream_head_t {
 #define IS_CLIENT_STREAM_ID(id) (unsigned int)(((id) & 1) == 0)
 #define IS_BIDIR_STREAM_ID(id)  (unsigned int)(((id) & 2) == 0)
 #define IS_LOCAL_STREAM_ID(id, client_mode)  (unsigned int)(((id)^(client_mode)) & 1)
-#define STREAM_ID_FROM_RANK(rank, is_server_stream, is_unidir) (((rank-1)<<2)|((is_unidir)<<1)|(is_server_stream))
+#define STREAM_ID_FROM_RANK(rank, is_server_stream, is_unidir) (((rank-1ull)<<2)|(uint64_t)((is_unidir)<<1)|(is_server_stream))
 #define STREAM_RANK_FROM_ID(id) ((id + 4)>>2)
 #define STREAM_TYPE_FROM_ID(id) ((id)&3)
 #define NEXT_STREAM_ID_FOR_TYPE(id) ((id)+4)
